+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## Enhancements 改进
+
+1. Show fetching status at footer play bar. 在播放底栏显示加载状态。
+
 # 1.1.0 (2026-01-26)
 
 ## Features 新功能